+++ packages/cesium-wind-layer/src/shaders/updatePosition.ts
@@ -10,13 +10,13 @@ out vec4 fragColor;
 
 void main() {
     // 获取当前粒子的位置
-    vec2 currentPos = texture(currentParticlesPosition, v_textureCoordinates).rg;
+    vec3 currentPos = texture(currentParticlesPosition, v_textureCoordinates).rgb;
     // 获取粒子的速度
     vec2 speed = texture(particlesSpeed, v_textureCoordinates).rg;
     // 计算下一个位置
-    vec2 nextPos = currentPos + speed;
+    vec3 nextPos = vec3(currentPos.xy + speed, currentPos.z);
     
     // 将新的位置写入 fragColor
-    fragColor = vec4(nextPos, 0.0, 1.0);
+    fragColor = vec4(nextPos, 1.0);
 }
 `;